Earth image
Instead of using map of earth, the new plan is to use a globe with the starfield animation behind it. Can also use this in the roadmap section, but that one should probably have a rocket and/or astronaut as well. I had to create this svg myself, since there wasn't a suitable one available